Module lwip_rust::bindings

Structs

__fsid_t

eth_addr

An Ethernet MAC address

eth_hdr

Ethernet header

eth_vlan_hdr

VLAN header inserted between ethernet header and payload if ‘type’ in ethernet header is ETHTYPE_VLAN. See IEEE802.Q

etharp_hdr

the ARP message, see RFC 826 (“Packet format”)

icmp6_echo_hdr

icmp6_hdr

icmp_echo_hdr

This is the standard ICMP header only that the u32_t data is split to two u16_t like ICMP echo needs it. This header is also used for other ICMP types that do not use the data part.

ip4_addr

This is the aligned version of ip4_addr_t, used as local variable, on the stack, etc.

ip4_addr_packed

ip4_addr_wordaligned

ip6_addr

This is the aligned version of ip6_addr_t, used as local variable, on the stack, etc.

ip6_addr_packed

ip6_dest_hdr

ip6_frag_hdr

ip6_hbh_hdr

ip6_hdr

ip6_opt_hdr

ip6_rout_hdr

ip_addr

@ingroup ipaddr A union struct for both IP version’s addresses. ATTENTION: watch out for its size when adding IPv6 address scope!

ip_globals

Global variables of this module, kept in a struct for efficient access using base+index.

ip_hdr

ip_pcb

lwip_cyclic_timer

This struct contains information about a stack-internal timer function that has to be called at a defined interval

max_align_t

memp

memp_desc

Memory pool descriptor

netif

Generic data structure used for all lwIP network interfaces. The following fields should be filled in by the initialization function for the device driver: hwaddr_len, hwaddr[], mtu, flags

netif_ext_callback_args_t_ipv4_changed_s

Args to LWIP_NSC_IPV4_ADDRESS_CHANGED|LWIP_NSC_IPV4_GATEWAY_CHANGED|LWIP_NSC_IPV4_NETMASK_CHANGED|LWIP_NSC_IPV4_SETTINGS_CHANGED callback

netif_ext_callback_args_t_ipv6_addr_state_changed_s

Args to LWIP_NSC_IPV6_ADDR_STATE_CHANGED callback

netif_ext_callback_args_t_ipv6_set_s

Args to LWIP_NSC_IPV6_SET callback

netif_ext_callback_args_t_link_changed_s

Args to LWIP_NSC_LINK_CHANGED callback

netif_ext_callback_args_t_status_changed_s

Args to LWIP_NSC_STATUS_CHANGED callback

pbuf

Main packet buffer struct

pbuf_custom

A custom pbuf: like a pbuf, but following a function pointer to free it.

pbuf_rom

Helper struct for const-correctness only. The only meaning of this one is to provide a const payload pointer for PBUF_ROM type.

rx_custom_pbuf_t

sys_timeo

tcp_ext_arg_callbacks

A table of callback functions that is invoked for ext arguments

tcp_pcb

the TCP protocol control block

tcp_pcb_listen

the TCP protocol control block for listening pcbs

tcp_seg

udp_hdr

udp_pcb

the UDP protocol control block

Constants

ARP_MAXAGE

ARP_QUEUEING

ARP_QUEUE_LEN

ARP_TABLE_SIZE

ARP_TMR_INTERVAL

BIG_ENDIAN

BYTE_ORDER

CHECKSUM_CHECK_ICMP

CHECKSUM_CHECK_ICMP6

CHECKSUM_CHECK_IP

CHECKSUM_CHECK_TCP

CHECKSUM_CHECK_UDP

CHECKSUM_GEN_ICMP

CHECKSUM_GEN_ICMP6

CHECKSUM_GEN_IP

CHECKSUM_GEN_TCP

CHECKSUM_GEN_UDP

DEFAULT_ACCEPTMBOX_SIZE

DEFAULT_RAW_RECVMBOX_SIZE

DEFAULT_TCP_RECVMBOX_SIZE

DEFAULT_THREAD_NAME

DEFAULT_THREAD_PRIO

DEFAULT_THREAD_STACKSIZE

DEFAULT_UDP_RECVMBOX_SIZE

DNS_DOES_NAME_CHECK

DNS_LOCAL_HOSTLIST

DNS_LOCAL_HOSTLIST_IS_DYNAMIC

DNS_MAX_NAME_LENGTH

DNS_MAX_RETRIES

DNS_MAX_SERVERS

DNS_TABLE_SIZE

DNS_TMR_INTERVAL

ETHARP_HWADDR_LEN

ETHARP_STATS

ETHARP_SUPPORT_STATIC_ENTRIES

ETHARP_SUPPORT_VLAN

ETH_HWADDR_LEN

ETH_PAD_SIZE

ICMP6_HLEN

ICMP6_STATS

ICMP_AM

ICMP_AMR

ICMP_DUR

ICMP_ECHO

ICMP_ER

ICMP_IR

ICMP_IRQ

ICMP_PP

ICMP_RD

ICMP_SQ

ICMP_STATS

ICMP_TE

ICMP_TS

ICMP_TSR

ICMP_TTL

IGMP_STATS

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

IP4ADDR_STRLEN_MAX

IP6ADDR_STRLEN_MAX

IP6_ADDR_DEPRECATED

IP6_ADDR_DUPLICATED

IP6_ADDR_INVALID

IP6_ADDR_LIFE_INFINITE

IP6_ADDR_LIFE_STATIC

IP6_ADDR_PREFERRED

IP6_ADDR_TENTATIVE

IP6_ADDR_TENTATIVE_1

IP6_ADDR_TENTATIVE_2

IP6_ADDR_TENTATIVE_3

IP6_ADDR_TENTATIVE_4

IP6_ADDR_TENTATIVE_5

IP6_ADDR_TENTATIVE_6

IP6_ADDR_TENTATIVE_7

IP6_ADDR_TENTATIVE_COUNT_MASK

IP6_ADDR_VALID

IP6_DEST_HLEN

IP6_FRAG_HLEN

IP6_FRAG_MORE_FLAG

IP6_FRAG_OFFSET_MASK

IP6_FRAG_STATS

IP6_HBH_HLEN

IP6_HLEN

IP6_HOME_ADDRESS_OPTION

IP6_JUMBO_OPTION

IP6_MIN_MTU_LENGTH

IP6_MULTICAST_SCOPE_ADMIN_LOCAL

IP6_MULTICAST_SCOPE_GLOBAL

IP6_MULTICAST_SCOPE_INTERFACE_LOCAL

IP6_MULTICAST_SCOPE_LINK_LOCAL

IP6_MULTICAST_SCOPE_ORGANIZATION_LOCAL

IP6_MULTICAST_SCOPE_RESERVED

IP6_MULTICAST_SCOPE_RESERVED0

IP6_MULTICAST_SCOPE_RESERVED3

IP6_MULTICAST_SCOPE_RESERVEDF

IP6_MULTICAST_SCOPE_SITE_LOCAL

IP6_NEXTH_DESTOPTS

IP6_NEXTH_ENCAPS

IP6_NEXTH_FRAGMENT

IP6_NEXTH_HOPBYHOP

IP6_NEXTH_ICMP6

IP6_NEXTH_NONE

IP6_NEXTH_ROUTING

IP6_NEXTH_TCP

IP6_NEXTH_UDP

IP6_NEXTH_UDPLITE

IP6_NO_ZONE

IP6_OPT_HLEN

IP6_PAD1_OPTION

IP6_PADN_OPTION

IP6_ROUTER_ALERT_DLEN

IP6_ROUTER_ALERT_OPTION

IP6_ROUTER_ALERT_VALUE_MLD

IP6_ROUT_RPL

IP6_ROUT_TYPE2

IP6_STATS

IPADDR_STRLEN_MAX

IPFRAG_STATS

IPV6_CUSTOM_SCOPES

IPV6_FRAG_COPYHEADER

IPV6_REASS_MAXAGE

IP_CLASSA_HOST

IP_CLASSA_MAX

IP_CLASSA_NET

IP_CLASSA_NSHIFT

IP_CLASSB_HOST

IP_CLASSB_MAX

IP_CLASSB_NET

IP_CLASSB_NSHIFT

IP_CLASSC_HOST

IP_CLASSC_NET

IP_CLASSC_NSHIFT

IP_CLASSD_HOST

IP_CLASSD_NET

IP_CLASSD_NSHIFT

IP_DEFAULT_TTL

IP_DF

IP_FORWARD

IP_FORWARD_ALLOW_TX_ON_RX_NETIF

IP_FRAG

IP_HLEN

IP_HLEN_MAX

IP_LOOPBACKNET

IP_MF

IP_OFFMASK

IP_OPTIONS_ALLOWED

IP_PROTO_ICMP

IP_PROTO_IGMP

IP_PROTO_TCP

IP_PROTO_UDP

IP_PROTO_UDPLITE

IP_REASSEMBLY

IP_REASS_MAXAGE

IP_REASS_MAX_PBUFS

IP_RF

IP_SOF_BROADCAST

IP_SOF_BROADCAST_RECV

IP_STATS

LINK_STATS

LITTLE_ENDIAN

LL_IP4_MULTICAST_ADDR_0

LL_IP4_MULTICAST_ADDR_1

LL_IP4_MULTICAST_ADDR_2

LL_IP6_MULTICAST_ADDR_0

LL_IP6_MULTICAST_ADDR_1

LWIP_ALLOW_MEM_FREE_FROM_OTHER_CONTEXT

LWIP_ALTCP

LWIP_ALTCP_TLS

LWIP_ARP

LWIP_ARP_FILTER_NETIF

LWIP_AUTOIP

LWIP_BROADCAST_PING

LWIP_CALLBACK_API

LWIP_CHECKSUM_CTRL_PER_NETIF

LWIP_CHECKSUM_ON_COPY

LWIP_COMPAT_SOCKETS

LWIP_DBG_FRESH

LWIP_DBG_HALT

LWIP_DBG_LEVEL_ALL

LWIP_DBG_LEVEL_OFF

LWIP_DBG_LEVEL_SERIOUS

LWIP_DBG_LEVEL_SEVERE

LWIP_DBG_LEVEL_WARNING

LWIP_DBG_MASK_LEVEL

LWIP_DBG_OFF

LWIP_DBG_ON

LWIP_DBG_STATE

LWIP_DBG_TRACE

LWIP_DEBUG

LWIP_DHCP

LWIP_DHCP6_GET_NTP_SRV

LWIP_DHCP6_MAX_DNS_SERVERS

LWIP_DHCP6_MAX_NTP_SERVERS

LWIP_DHCP_AUTOIP_COOP

LWIP_DHCP_AUTOIP_COOP_TRIES

LWIP_DHCP_BOOTP_FILE

LWIP_DHCP_GET_NTP_SRV

LWIP_DHCP_MAX_NTP_SERVERS

LWIP_DNS

LWIP_DNS_ADDRTYPE_DEFAULT

LWIP_DNS_ADDRTYPE_IPV4

LWIP_DNS_ADDRTYPE_IPV4_IPV6

LWIP_DNS_ADDRTYPE_IPV6

LWIP_DNS_ADDRTYPE_IPV6_IPV4

LWIP_DNS_SECURE_NO_MULTIPLE_OUTSTANDING

LWIP_DNS_SECURE_RAND_SRC_PORT

LWIP_DNS_SECURE_RAND_XID

LWIP_DNS_SUPPORT_MDNS_QUERIES

LWIP_ETHERNET

LWIP_EVENT_API

LWIP_FIONREAD_LINUXMODE

LWIP_HAVE_INT64

LWIP_HAVE_LOOPIF

LWIP_HAVE_SLIPIF

LWIP_ICMP

LWIP_ICMP6

LWIP_ICMP6_DATASIZE

LWIP_ICMP6_HL

LWIP_IGMP

LWIP_IPV4

LWIP_IPV4_SRC_ROUTING

LWIP_IPV6

LWIP_IPV6_ADDRESS_LIFETIMES

LWIP_IPV6_AUTOCONFIG

LWIP_IPV6_DHCP6

LWIP_IPV6_DHCP6_STATEFUL

LWIP_IPV6_DHCP6_STATELESS

LWIP_IPV6_DUP_DETECT_ATTEMPTS

LWIP_IPV6_FORWARD

LWIP_IPV6_FRAG

LWIP_IPV6_MLD

LWIP_IPV6_NUM_ADDRESSES

LWIP_IPV6_REASS

LWIP_IPV6_SCOPES_DEBUG

LWIP_IPV6_SEND_ROUTER_SOLICIT

LWIP_LOOPBACK_MAX_PBUFS

LWIP_LOOPIF_MULTICAST

LWIP_MIB2_CALLBACKS

LWIP_MPU_COMPATIBLE

LWIP_MULTICAST_PING

LWIP_ND6_ALLOW_RA_UPDATES

LWIP_ND6_DELAY_FIRST_PROBE_TIME

LWIP_ND6_MAX_ANYCAST_DELAY_TIME

LWIP_ND6_MAX_MULTICAST_SOLICIT

LWIP_ND6_MAX_NEIGHBOR_ADVERTISEMENT

LWIP_ND6_MAX_UNICAST_SOLICIT

LWIP_ND6_NUM_DESTINATIONS

LWIP_ND6_NUM_NEIGHBORS

LWIP_ND6_NUM_PREFIXES

LWIP_ND6_NUM_ROUTERS

LWIP_ND6_QUEUEING

LWIP_ND6_RDNSS_MAX_DNS_SERVERS

LWIP_ND6_REACHABLE_TIME

LWIP_ND6_RETRANS_TIMER

LWIP_ND6_TCP_REACHABILITY_HINTS

LWIP_NETBUF_RECVINFO

LWIP_NETCONN

LWIP_NETCONN_FULLDUPLEX

LWIP_NETCONN_SEM_PER_THREAD

LWIP_NETIF_API

LWIP_NETIF_EXT_STATUS_CALLBACK

LWIP_NETIF_HOSTNAME

LWIP_NETIF_HWADDRHINT

LWIP_NETIF_LINK_CALLBACK

LWIP_NETIF_LOOPBACK

LWIP_NETIF_REMOVE_CALLBACK

LWIP_NETIF_STATUS_CALLBACK

LWIP_NETIF_TX_SINGLE_PBUF

LWIP_NETIF_USE_HINTS

LWIP_NO_CTYPE_H

LWIP_NO_INTTYPES_H

LWIP_NO_LIMITS_H

LWIP_NO_STDDEF_H

LWIP_NO_STDINT_H

LWIP_NO_UNISTD_H

LWIP_NSC_IPV4_ADDRESS_CHANGED

LWIP_NSC_IPV4_GATEWAY_CHANGED

LWIP_NSC_IPV4_NETMASK_CHANGED

LWIP_NSC_IPV4_SETTINGS_CHANGED

LWIP_NSC_IPV6_ADDR_STATE_CHANGED

LWIP_NSC_IPV6_SET

LWIP_NSC_LINK_CHANGED

LWIP_NSC_NETIF_ADDED

LWIP_NSC_NETIF_REMOVED

LWIP_NSC_NONE

LWIP_NSC_STATUS_CHANGED

LWIP_NUM_NETIF_CLIENT_DATA

LWIP_PERF

LWIP_POSIX_SOCKETS_IO_NAMES

LWIP_RAW

LWIP_RC_DEVELOPMENT

LWIP_RC_RELEASE

LWIP_SINGLE_NETIF

LWIP_SNMP

LWIP_SOCKET

LWIP_SOCKET_OFFSET

LWIP_SOCKET_POLL

LWIP_SOCKET_SELECT

LWIP_SO_LINGER

LWIP_SO_RCVBUF

LWIP_SO_RCVTIMEO

LWIP_SO_SNDRCVTIMEO_NONSTANDARD

LWIP_SO_SNDTIMEO

LWIP_STATS

LWIP_STATS_DISPLAY

LWIP_TCP

LWIP_TCPIP_CORE_LOCKING

LWIP_TCPIP_CORE_LOCKING_INPUT

LWIP_TCPIP_TIMEOUT

LWIP_TCP_CLOSE_TIMEOUT_MS_DEFAULT

LWIP_TCP_KEEPALIVE

LWIP_TCP_MAX_SACK_NUM

LWIP_TCP_PCB_NUM_EXT_ARGS

LWIP_TCP_PCB_NUM_EXT_ARG_ID_INVALID

LWIP_TCP_SACK_OUT

LWIP_TCP_TIMESTAMPS

LWIP_TESTMODE

LWIP_TIMERS

LWIP_TIMERS_CUSTOM

LWIP_UDP

LWIP_UDPLITE

LWIP_UINT32_MAX

LWIP_VERSION_MAJOR

LWIP_VERSION_MINOR

LWIP_VERSION_REVISION

LWIP_VERSION_STRING_SUFFIX

LWIP_WND_SCALE

MEMP_MEM_INIT

MEMP_MEM_MALLOC

MEMP_NUM_ALTCP_PCB

MEMP_NUM_API_MSG

MEMP_NUM_ARP_QUEUE

MEMP_NUM_DNS_API_MSG

MEMP_NUM_FRAG_PBUF

MEMP_NUM_IGMP_GROUP

MEMP_NUM_LOCALHOSTLIST

MEMP_NUM_MLD6_GROUP

MEMP_NUM_ND6_QUEUE

MEMP_NUM_NETBUF

MEMP_NUM_NETCONN

MEMP_NUM_NETDB

MEMP_NUM_NETIFAPI_MSG

MEMP_NUM_PBUF

MEMP_NUM_RAW_PCB

MEMP_NUM_REASSDATA

MEMP_NUM_SELECT_CB

MEMP_NUM_SOCKET_SETGETSOCKOPT_DATA

MEMP_NUM_TCPIP_MSG_API

MEMP_NUM_TCPIP_MSG_INPKT

MEMP_NUM_TCP_PCB

MEMP_NUM_TCP_PCB_LISTEN

MEMP_NUM_TCP_SEG

MEMP_NUM_UDP_PCB

MEMP_OVERFLOW_CHECK

MEMP_SANITY_CHECK

MEMP_SIZE

MEMP_STATS

MEMP_USE_CUSTOM_POOLS

MEM_ALIGNMENT

MEM_LIBC_MALLOC

MEM_OVERFLOW_CHECK

MEM_SANITY_CHECK

MEM_SIZE

MEM_SIZE_F

MEM_STATS

MEM_USE_POOLS

MEM_USE_POOLS_TRY_BIGGER_POOL

MIB2_STATS

MLD6_STATS

ND6_STATS

NETIF_ADDR_IDX_MAX

NETIF_FLAG_BROADCAST

NETIF_FLAG_ETHARP

NETIF_FLAG_ETHERNET

NETIF_FLAG_IGMP

NETIF_FLAG_LINK_UP

NETIF_FLAG_MLD6

NETIF_FLAG_UP

NETIF_MAX_HWADDR_LEN

NETIF_NAMESIZE

NETIF_NO_INDEX

NO_SYS

PBUF_ALLOC_FLAG_DATA_CONTIGUOUS

PBUF_ALLOC_FLAG_RX

PBUF_FLAG_IS_CUSTOM

PBUF_FLAG_LLBCAST

PBUF_FLAG_LLMCAST

PBUF_FLAG_MCASTLOOP

PBUF_FLAG_PUSH

PBUF_FLAG_TCP_FIN

PBUF_IP_HLEN

PBUF_LINK_ENCAPSULATION_HLEN

PBUF_LINK_HLEN

PBUF_POOL_FREE_OOSEQ

PBUF_POOL_SIZE

PBUF_TRANSPORT_HLEN

PBUF_TYPE_ALLOC_SRC_MASK

PBUF_TYPE_ALLOC_SRC_MASK_APP_MAX

PBUF_TYPE_ALLOC_SRC_MASK_APP_MIN

PBUF_TYPE_ALLOC_SRC_MASK_STD_HEAP

PBUF_TYPE_ALLOC_SRC_MASK_STD_MEMP_PBUF

PBUF_TYPE_ALLOC_SRC_MASK_STD_MEMP_PBUF_POOL

PBUF_TYPE_FLAG_DATA_VOLATILE

PBUF_TYPE_FLAG_STRUCT_DATA_CONTIGUOUS

PPP_SUPPORT

PTRDIFF_MAX

PTRDIFF_MIN

RAW_TTL

S8_F

S16_F

S32_F

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZEOF_ETHARP_HDR

SIZEOF_ETH_HDR

SIZEOF_VLAN_HDR

SIZE_MAX

SLIPIF_THREAD_NAME

SLIPIF_THREAD_PRIO

SLIPIF_THREAD_STACKSIZE

SOF_BROADCAST

SOF_INHERITED

SOF_KEEPALIVE

SOF_REUSEADDR

SO_REUSE

SO_REUSE_RXTOALL

SYS_LIGHTWEIGHT_PROT

SYS_STATS

SYS_TIMEOUTS_SLEEPTIME_INFINITE

SZT_F

TCPIP_MBOX_SIZE

TCPIP_THREAD_NAME

TCPIP_THREAD_PRIO

TCPIP_THREAD_STACKSIZE

TCP_ALLFLAGS

TCP_CALCULATE_EFF_SEND_MSS

TCP_DEFAULT_LISTEN_BACKLOG

TCP_LISTEN_BACKLOG

TCP_MAXRTX

TCP_MSS

TCP_OOSEQ_MAX_BYTES

TCP_OOSEQ_MAX_PBUFS

TCP_OVERSIZE

TCP_PRIO_MAX

TCP_PRIO_MIN

TCP_PRIO_NORMAL

TCP_QUEUE_OOSEQ

TCP_RCV_SCALE

TCP_SNDQUEUELEN_OVERFLOW

TCP_SND_BUF

TCP_SND_QUEUELEN

TCP_STATS

TCP_SYNMAXRTX

TCP_TTL

TCP_WND

TCP_WRITE_FLAG_COPY

TCP_WRITE_FLAG_MORE

TF_ACK_DELAY

TF_ACK_NOW

TF_CLOSEPEND

TF_FIN

TF_INFR

TF_NAGLEMEMERR

TF_NODELAY

TF_RTO

TF_RXCLOSED

U8_F

U16_F

U32_F

UDP_FLAGS_CONNECTED

UDP_FLAGS_MULTICAST_LOOP

UDP_FLAGS_NOCHKSUM

UDP_FLAGS_UDPLITE

UDP_HLEN

UDP_STATS

UDP_TTL

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

X8_F

X16_F

X32_F

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

err_enum_t_ERR_ABRT

Connection aborted.

err_enum_t_ERR_ALREADY

Already connecting.

err_enum_t_ERR_ARG

Illegal argument.

err_enum_t_ERR_BUF

Buffer error.

err_enum_t_ERR_CLSD

Connection closed.

err_enum_t_ERR_CONN

Not connected.

err_enum_t_ERR_IF

Low-level netif error

err_enum_t_ERR_INPROGRESS

Operation in progress

err_enum_t_ERR_ISCONN

Conn already established.

err_enum_t_ERR_MEM

Out of memory error.

err_enum_t_ERR_OK

No error, everything OK.

err_enum_t_ERR_RST

Connection reset.

err_enum_t_ERR_RTE

Routing problem.

err_enum_t_ERR_TIMEOUT

Timeout.

err_enum_t_ERR_USE

Address in use.

err_enum_t_ERR_VAL

Illegal value.

err_enum_t_ERR_WOULDBLOCK

Operation would block.

etharp_opcode_ARP_REPLY

etharp_opcode_ARP_REQUEST

icmp6_dur_code_ICMP6_DUR_ADDRESS

Address unreachable

icmp6_dur_code_ICMP6_DUR_NO_ROUTE

No route to destination

icmp6_dur_code_ICMP6_DUR_POLICY

Source address failed ingress/egress policy

icmp6_dur_code_ICMP6_DUR_PORT

Port unreachable

icmp6_dur_code_ICMP6_DUR_PROHIBITED

Communication with destination administratively prohibited

icmp6_dur_code_ICMP6_DUR_REJECT_ROUTE

Reject route to destination

icmp6_dur_code_ICMP6_DUR_SCOPE

Beyond scope of source address

icmp6_pp_code_ICMP6_PP_FIELD

Erroneous header field encountered

icmp6_pp_code_ICMP6_PP_HEADER

Unrecognized next header type encountered

icmp6_pp_code_ICMP6_PP_OPTION

Unrecognized IPv6 option encountered

icmp6_te_code_ICMP6_TE_FRAG

Fragment reassembly time exceeded

icmp6_te_code_ICMP6_TE_HL

Hop limit exceeded in transit

icmp6_type_ICMP6_TYPE_DUR

Destination unreachable

icmp6_type_ICMP6_TYPE_EREP

Echo reply

icmp6_type_ICMP6_TYPE_EREQ

Echo request

icmp6_type_ICMP6_TYPE_MLD

Multicast listener done

icmp6_type_ICMP6_TYPE_MLQ

Multicast listener query

icmp6_type_ICMP6_TYPE_MLR

Multicast listener report

icmp6_type_ICMP6_TYPE_MRA

Multicast router advertisement

icmp6_type_ICMP6_TYPE_MRS

Multicast router solicitation

icmp6_type_ICMP6_TYPE_MRT

Multicast router termination

icmp6_type_ICMP6_TYPE_NA

Neighbor advertisement

icmp6_type_ICMP6_TYPE_NS

Neighbor solicitation

icmp6_type_ICMP6_TYPE_PE1

Private experimentation

icmp6_type_ICMP6_TYPE_PE2

Private experimentation

icmp6_type_ICMP6_TYPE_PE3

Private experimentation

icmp6_type_ICMP6_TYPE_PE4

Private experimentation

icmp6_type_ICMP6_TYPE_PP

Parameter problem

icmp6_type_ICMP6_TYPE_PTB

Packet too big

icmp6_type_ICMP6_TYPE_RA

Router advertisement

icmp6_type_ICMP6_TYPE_RD

Redirect

icmp6_type_ICMP6_TYPE_RS

Router solicitation

icmp6_type_ICMP6_TYPE_RSV_ERR

Reserved for expansion of error messages

icmp6_type_ICMP6_TYPE_RSV_INF

Reserved for expansion of informational messages

icmp6_type_ICMP6_TYPE_TE

Time exceeded

icmp_dur_type_ICMP_DUR_FRAG

fragmentation needed and DF set

icmp_dur_type_ICMP_DUR_HOST

host unreachable

icmp_dur_type_ICMP_DUR_NET

net unreachable

icmp_dur_type_ICMP_DUR_PORT

port unreachable

icmp_dur_type_ICMP_DUR_PROTO

protocol unreachable

icmp_dur_type_ICMP_DUR_SR

source route failed

icmp_te_type_ICMP_TE_FRAG

fragment reassembly time exceeded

icmp_te_type_ICMP_TE_TTL

time to live exceeded in transit

lwip_ieee_eth_type_ETHTYPE_ARP

Address resolution protocol

lwip_ieee_eth_type_ETHTYPE_ETHERCAT

Ethernet for control automation technology

lwip_ieee_eth_type_ETHTYPE_IP

Internet protocol v4

lwip_ieee_eth_type_ETHTYPE_IPV6

Internet protocol v6

lwip_ieee_eth_type_ETHTYPE_JUMBO

Jumbo Frames

lwip_ieee_eth_type_ETHTYPE_LLDP

Link layer discovery protocol

lwip_ieee_eth_type_ETHTYPE_MRP

Media redundancy protocol

lwip_ieee_eth_type_ETHTYPE_PPPOE

PPP Over Ethernet Session Stage

lwip_ieee_eth_type_ETHTYPE_PPPOEDISC

PPP Over Ethernet Discovery Stage

lwip_ieee_eth_type_ETHTYPE_PROFINET

Process field network

lwip_ieee_eth_type_ETHTYPE_PTP

Precision time protocol

lwip_ieee_eth_type_ETHTYPE_QINQ

Q-in-Q, 802.1ad

lwip_ieee_eth_type_ETHTYPE_RARP

RARP

lwip_ieee_eth_type_ETHTYPE_SERCOS

Serial real-time communication system

lwip_ieee_eth_type_ETHTYPE_VLAN

Virtual local area network

lwip_ieee_eth_type_ETHTYPE_WOL

Wake on lan

lwip_internal_netif_client_data_index_LWIP_NETIF_CLIENT_DATA_INDEX_MAX

lwip_ip_addr_type_IPADDR_TYPE_ANY

IPv4+IPv6 (“dual-stack”)

lwip_ip_addr_type_IPADDR_TYPE_V4

IPv4

lwip_ip_addr_type_IPADDR_TYPE_V6

IPv6

lwip_ipv6_scope_type_IP6_MULTICAST

Multicast

lwip_ipv6_scope_type_IP6_UNICAST

Unicast

lwip_ipv6_scope_type_IP6_UNKNOWN

Unknown

memp_t_MEMP_FRAG_PBUF

memp_t_MEMP_IP6_REASSDATA

memp_t_MEMP_MAX

memp_t_MEMP_ND6_QUEUE

memp_t_MEMP_PBUF

memp_t_MEMP_PBUF_POOL

memp_t_MEMP_REASSDATA

memp_t_MEMP_SYS_TIMEOUT

memp_t_MEMP_TCP_PCB

memp_t_MEMP_TCP_PCB_LISTEN

memp_t_MEMP_TCP_SEG

memp_t_MEMP_UDP_PCB

netif_mac_filter_action_NETIF_ADD_MAC_FILTER

Add a filter entry

netif_mac_filter_action_NETIF_DEL_MAC_FILTER

Delete a filter entry

pbuf_layer_PBUF_IP

Includes spare room for IP header. Use this if you intend to pass the pbuf to functions like raw_send().

pbuf_layer_PBUF_LINK

Includes spare room for link layer header (ethernet header). Use this if you intend to pass the pbuf to functions like ethernet_output(). @see PBUF_LINK_HLEN

pbuf_layer_PBUF_RAW

Use this for input packets in a netif driver when calling netif->input() in the most common case - ethernet-layer netif driver.

pbuf_layer_PBUF_RAW_TX

Includes spare room for additional encapsulation header before ethernet headers (e.g. 802.11). Use this if you intend to pass the pbuf to functions like netif->linkoutput(). @see PBUF_LINK_ENCAPSULATION_HLEN

pbuf_layer_PBUF_TRANSPORT

Includes spare room for transport layer header, e.g. UDP header. Use this if you intend to pass the pbuf to functions like udp_send().

pbuf_type_PBUF_POOL

pbuf payload refers to RAM. This one comes from a pool and should be used for RX. Payload can be chained (scatter-gather RX) but like PBUF_RAM, struct pbuf and its payload are allocated in one piece of contiguous memory (so the first payload byte can be calculated from struct pbuf). Don’t use this for TX, if the pool becomes empty e.g. because of TCP queuing, you are unable to receive TCP acks!

pbuf_type_PBUF_RAM

pbuf data is stored in RAM, used for TX mostly, struct pbuf and its payload are allocated in one piece of contiguous memory (so the first payload byte can be calculated from struct pbuf). pbuf_alloc() allocates PBUF_RAM pbufs as unchained pbufs (although that might change in future versions). This should be used for all OUTGOING packets (TX).

pbuf_type_PBUF_REF

pbuf comes from the pbuf pool. Much like PBUF_ROM but payload might change so it has to be duplicated when queued before transmitting, depending on who has a ‘ref’ to it.

pbuf_type_PBUF_ROM

pbuf data is stored in ROM, i.e. struct pbuf and its payload are located in totally different memory areas. Since it points to ROM, payload does not have to be copied when queued for transmission.

tcp_state_CLOSED

tcp_state_CLOSE_WAIT

tcp_state_CLOSING

tcp_state_ESTABLISHED

tcp_state_FIN_WAIT_1

tcp_state_FIN_WAIT_2

tcp_state_LAST_ACK

tcp_state_LISTEN

tcp_state_SYN_RCVD

tcp_state_SYN_SENT

tcp_state_TIME_WAIT

Statics

dns_mquery_v4group

dns_mquery_v6group

ethbroadcast

ethzero

ip6_addr_any

ip_addr_any

ip_addr_any_type

ip_addr_broadcast

ip_data

lwip_cyclic_timers

This array contains all stack-internal cyclic timers. To get the number of timers, use lwip_num_cyclic_timers

lwip_num_cyclic_timers

Array size of lwip_cyclic_timers[]

memp_pools

netif_default

The default network interface.

netif_list

The list of network interfaces.

pbuf_free_ooseq_pending

udp_pcbs

Functions

dns_gethostbyname⚠

dns_gethostbyname_addrtype⚠

dns_getserver⚠

dns_init⚠

dns_setserver⚠

dns_tmr⚠

etharp_cleanup_netif⚠

etharp_find_addr⚠

etharp_get_entry⚠

etharp_input⚠

etharp_output⚠

etharp_query⚠

etharp_request⚠

etharp_tmr⚠

ethernet_input⚠

ethernet_output⚠

ethip6_output⚠

icmp6_dest_unreach⚠

icmp6_input⚠

icmp6_packet_too_big⚠

icmp6_param_problem⚠

icmp6_time_exceeded⚠

icmp6_time_exceeded_with_addrs⚠

icmp_dest_unreach⚠

icmp_input⚠

icmp_time_exceeded⚠

ip4_addr_isbroadcast_u32⚠

ip4_addr_netmask_valid⚠

ip4_debug_print⚠

ip4_input⚠

ip4_output⚠

ip4_output_if⚠

ip4_output_if_src⚠

ip4_route⚠

ip4addr_aton⚠

ip4addr_ntoa⚠

returns ptr to static buffer; not reentrant!

ip4addr_ntoa_r⚠

ip6_debug_print⚠

ip6_input⚠

ip6_output⚠

ip6_output_if⚠

ip6_output_if_src⚠

ip6_route⚠

ip6_select_source_address⚠

ip6addr_aton⚠

ip6addr_ntoa⚠

returns ptr to static buffer; not reentrant!

ip6addr_ntoa_r⚠

ip_input⚠

ipaddr_addr⚠

ipaddr_aton⚠

ipaddr_ntoa⚠

ipaddr_ntoa_r⚠

lwip_abort⚠

lwip_htonl⚠

lwip_htons⚠

lwip_init⚠

@}

lwip_itoa⚠

lwip_print⚠

lwip_rand_u32⚠

lwip_strerr⚠

lwip_stricmp⚠

lwip_strnicmp⚠

lwip_strnstr⚠

mem_calloc⚠

mem_free⚠

mem_init⚠

mem_malloc⚠

mem_trim⚠

memp_free⚠

memp_free_pool⚠

memp_init⚠

memp_init_pool⚠

memp_malloc⚠

memp_malloc_pool⚠

netif_add⚠

netif_add_ip6_address⚠

netif_add_noaddr⚠

netif_create_ip6_linklocal_address⚠

netif_find⚠

netif_get_by_index⚠

netif_get_ip6_addr_match⚠

netif_index_to_name⚠

netif_init⚠

netif_input⚠

netif_ip6_addr_set⚠

netif_ip6_addr_set_parts⚠

netif_ip6_addr_set_state⚠

netif_loop_output⚠

netif_name_to_index⚠

netif_poll⚠

netif_poll_all⚠

netif_remove⚠

netif_set_addr⚠

netif_set_default⚠

netif_set_down⚠

netif_set_gw⚠

netif_set_ipaddr⚠

netif_set_link_down⚠

netif_set_link_up⚠

netif_set_netmask⚠

netif_set_up⚠

pbuf_add_header⚠

pbuf_add_header_force⚠

pbuf_alloc⚠

pbuf_alloc_reference⚠

pbuf_alloced_custom⚠

pbuf_cat⚠

pbuf_chain⚠

pbuf_clen⚠

pbuf_clone⚠

pbuf_coalesce⚠

pbuf_copy⚠

pbuf_copy_partial⚠

pbuf_copy_partial_pbuf⚠

pbuf_dechain⚠

pbuf_fill_chksum⚠

pbuf_free⚠

pbuf_free_header⚠

pbuf_free_ooseq⚠

pbuf_get_at⚠

pbuf_get_contiguous⚠

pbuf_header⚠

pbuf_header_force⚠

pbuf_memcmp⚠

pbuf_memfind⚠

pbuf_put_at⚠

pbuf_realloc⚠

pbuf_ref⚠

pbuf_remove_header⚠

pbuf_skip⚠

pbuf_strstr⚠

pbuf_take⚠

pbuf_take_at⚠

pbuf_try_get_at⚠

rx_custom_pbuf_alloc⚠

rx_custom_pbuf_free⚠

rx_custom_pbuf_init⚠

sys_check_timeouts⚠

sys_restart_timeouts⚠

sys_timeout_debug⚠

sys_timeouts_init⚠

sys_timeouts_sleeptime⚠

sys_untimeout⚠

tcp_abort⚠

tcp_accept⚠

tcp_arg⚠

tcp_bind⚠

tcp_bind_netif⚠

tcp_close⚠

tcp_connect⚠

tcp_debug_state_str⚠

tcp_err⚠

tcp_listen_with_backlog⚠

tcp_listen_with_backlog_and_err⚠

tcp_new⚠

tcp_new_ip_type⚠

tcp_output⚠

tcp_poll⚠

tcp_recv⚠

tcp_recved⚠

tcp_sent⚠

tcp_setprio⚠

tcp_shutdown⚠

tcp_tcp_get_tcp_addrinfo⚠

tcp_write⚠

udp_bind⚠

udp_bind_netif⚠

udp_connect⚠

udp_debug_print⚠

udp_disconnect⚠

udp_init⚠

udp_input⚠

udp_netif_ip_addr_changed⚠

udp_new⚠

udp_new_ip_type⚠

udp_recv⚠

udp_remove⚠

udp_send⚠

udp_send_chksum⚠

udp_sendto⚠

udp_sendto_chksum⚠

udp_sendto_if⚠

udp_sendto_if_chksum⚠

udp_sendto_if_src⚠

udp_sendto_if_src_chksum⚠

Type Definitions

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

dns_found_callback

Callback which is invoked when a hostname is found. A function of this type must be implemented by the application using the DNS resolver. @param name pointer to the name that was looked up. @param ipaddr pointer to an ip_addr_t containing the IP address of the hostname, or NULL if the name could not be found (or on any other error). @param callback_arg a user-specified callback argument passed to dns_gethostbyname

err_enum_t

Definitions for error constants.

err_t

etharp_opcode

icmp6_dur_code

ICMP destination unreachable codes

icmp6_pp_code

ICMP parameter code

icmp6_te_code

ICMP time exceeded codes

icmp6_type

ICMP type

icmp_dur_type

ICMP destination unreachable codes

icmp_te_type

ICMP time exceeded codes

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

ip4_addr_p_t

ip4_addr_t

ip4_addr_t uses a struct for convenience only, so that the same defines can operate both on ip4_addr_t as well as on ip4_addr_p_t.

ip6_addr_p_t

ip6_addr_t

IPv6 address

ip_addr_t

@ingroup ipaddr A union struct for both IP version’s addresses. ATTENTION: watch out for its size when adding IPv6 address scope!

lwip_cyclic_timer_handler

Function prototype for a stack-internal timer function that has to be called at a defined interval

lwip_ieee_eth_type

@ingroup ieee A list of often ethtypes (although lwIP does not use all of them).

lwip_internal_netif_client_data_index

@}

lwip_ip_addr_type

@ingroup ipaddr IP address types for use in ip_addr_t.type member. @see tcp_new_ip_type(), udp_new_ip_type(), raw_new_ip_type().

lwip_ipv6_scope_type

Symbolic constants for the ‘type’ parameters in some of the macros. These exist for efficiency only, allowing the macros to avoid certain tests when the address is known not to be of a certain type. Dead code elimination will do the rest. IP6_MULTICAST is supported but currently not optimized. @see ip6_addr_has_scope, ip6_addr_assign_zone, ip6_addr_lacks_zone.

mem_ptr_t

mem_size_t

memp_t

Create the list of all memory pools managed by memp. MEMP_MAX represents a NULL pool at the end

netif_addr_idx_t

netif_ext_callback_fn

@ingroup netif Function used for extended netif status callbacks Note: When parsing reason argument, keep in mind that more reasons may be added in the future! @param netif netif that is affected by change @param reason change reason @param args depends on reason, see reason description

netif_init_fn

Function prototype for netif init functions. Set up flags and output/linkoutput callback functions in this function.

netif_input_fn

Function prototype for netif->input functions. This function is saved as ‘input’ callback function in the netif struct. Call it when a packet has been received.

netif_linkoutput_fn

Function prototype for netif->linkoutput functions. Only used for ethernet netifs. This function is called by ARP when a packet shall be sent.

netif_mac_filter_action

MAC Filter Actions, these are passed to a netif’s igmp_mac_filter or mld_mac_filter callback function.

netif_nsc_reason_t

@ingroup netif Extended netif status callback (NSC) reasons flags. May be extended in the future!

netif_output_fn

Function prototype for netif->output functions. Called by lwIP when a packet shall be sent. For ethernet netif, set this to ‘etharp_output’ and set ‘linkoutput’.

netif_output_ip6_fn

Function prototype for netif->output_ip6 functions. Called by lwIP when a packet shall be sent. For ethernet netif, set this to ‘ethip6_output’ and set ‘linkoutput’.

netif_status_callback_fn

Function prototype for netif status- or link-callback functions.

pbuf_free_custom_fn

Prototype for a function to free a custom pbuf

pbuf_layer

@ingroup pbuf Enumeration of pbuf layers

pbuf_type

@ingroup pbuf Enumeration of pbuf types

s8_t

s16_t

s32_t

s64_t

sys_timeout_handler

Function prototype for a timeout callback function. Register such a function using sys_timeout().

tcp_accept_fn

Function prototype for tcp accept callback functions. Called when a new connection can be accepted on a listening pcb.

tcp_connected_fn

Function prototype for tcp connected callback functions. Called when a pcb is connected to the remote side after initiating a connection attempt by calling tcp_connect().

tcp_err_fn

Function prototype for tcp error callback functions. Called when the pcb receives a RST or is unexpectedly closed for any other reason.

tcp_extarg_callback_passive_open_fn

Function prototype to transition arguments from a listening pcb to an accepted pcb

tcp_extarg_callback_pcb_destroyed_fn

Function prototype for deallocation of arguments. Called just before the pcb is freed, so don’t expect to be able to do anything with this pcb!

tcp_poll_fn

Function prototype for tcp poll callback functions. Called periodically as specified by @see tcp_poll.

tcp_recv_fn

Function prototype for tcp receive callback functions. Called when data has been received.

tcp_sent_fn

Function prototype for tcp sent callback functions. Called when sent data has been acknowledged by the remote side. Use it to free corresponding resources. This also means that the pcb has now space available to send new data.

tcp_state

tcpflags_t

tcpwnd_size_t

u8_t

u16_t

u32_t

u64_t

udp_recv_fn

Function prototype for udp pcb receive callback functions addr and port are in same byte order as in the pcb The callback is responsible for freeing the pbuf if it’s not used any more.

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t

Unions

ip_addr__bindgen_ty_1

netif_ext_callback_args_t

@ingroup netif Argument supplied to netif_ext_callback_fn.