Struct lwip_rust::bindings::ip_globals

#[repr(C)]
pub struct ip_globals {
    pub current_netif: *mut netif,
    pub current_input_netif: *mut netif,
    pub current_ip4_header: *const ip_hdr,
    pub current_ip6_header: *mut ip6_hdr,
    pub current_ip_header_tot_len: u16_t,
    pub current_iphdr_src: ip_addr_t,
    pub current_iphdr_dest: ip_addr_t,
}

Global variables of this module, kept in a struct for efficient access using base+index.

Fields
current_netif: *mut netif

The interface that accepted the packet for the current callback invocation.

current_input_netif: *mut netif

The interface that received the packet for the current callback invocation.

current_ip4_header: *const ip_hdr

Header of the input packet currently being processed.

current_ip6_header: *mut ip6_hdr

Header of the input IPv6 packet currently being processed.

current_ip_header_tot_len: u16_t

Total header length of current_ip4/6_header (i.e. after this, the UDP/TCP header starts)

current_iphdr_src: ip_addr_t

Source IP address of current_header

current_iphdr_dest: ip_addr_t

Destination IP address of current_header